chore(token): move responsiveStyle into token
