+++ packages/tokens/src/layout/index.ts
@@ -1,3 +1,4 @@
 export * from './breakpoints';
 export * from './grid';
+export * from './responsiveStyle';
 export * from './spacing';
